Split more rules
